refactor(core) Remove use of worker in collection service
